pyproject.toml
[tool.poetry]
name = "repl_python3_ChandanTestBot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
discord = "^1.7.3"
Flask = "^2.0.1"
helpers = "^0.2.0"
replit = "^3.2.4"
pytz = "^2021.1"
clr = "^1.0.3"
discord-py-interactions = "^3.0.2"
pythonnet = "^2.5.2"
clr-loader = "^0.1.6"
convertapi = "^1.4.0"
"dislash.py" = "^1.4.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
